+++ pds_github_util/utils/__init__.py
@@ -1,0 +1,2 @@
+from .githubConnection import GithubConnection
+from .RstClothReferenceable import RstClothReferenceable
